requirements.txt
attrs==19.3.0
click==7.0
flask==1.1.1
httpx==0.11.1
ipdb==0.13.3
pdfminer-six==20181108
python-magic==0.4.15
scrapy==2.2.0
